test(unit): refactore unit tests helper modules
